Split yaePlaylistView into 35 smaller files
