+++ apprenticevideo/yaeRectangle.cpp
@@ -1,0 +1,106 @@
+// -*- Mode: c++; tab-width: 8; c-basic-offset: 2; indent-tabs-mode: nil -*-
+// NOTE: the first line of this file sets up source code indentation rules
+// for Emacs; it is also a hint to anyone modifying this file.
+
+// Created      : Tue Oct 20 19:19:59 PDT 2015
+// Copyright    : Pavel Koshevoy
+// License      : MIT -- http://www.opensource.org/licenses/mit-license.php
+
+// Qt library:
+#include <QImage>
+
+// local interfaces:
+#include "yaeCanvasRenderer.h"
+#include "yaeRectangle.h"
+
+
+namespace yae
+{
+  //----------------------------------------------------------------
+  // Rectangle::Rectangle
+  //
+  Rectangle::Rectangle(const char * id):
+    Item(id),
+    border_(ItemRef::constant(0.0)),
+    color_(ColorRef::constant(Color(0x7f7f7f, 0.5))),
+    colorBorder_(ColorRef::constant(Color(0xffffff, 0.25)))
+  {}
+
+  //----------------------------------------------------------------
+  // paintRect
+  //
+  static void
+  paintRect(const BBox & bbox,
+            double border,
+            const Color & color,
+            const Color & colorBorder)
+  {
+    double x0 = bbox.x_;
+    double y0 = bbox.y_;
+    double x1 = bbox.w_ + x0;
+    double y1 = bbox.h_ + y0;
+
+    YAE_OGL_11_HERE();
+    YAE_OGL_11(glColor4ub(color.r(),
+                          color.g(),
+                          color.b(),
+                          color.a()));
+    YAE_OGL_11(glBegin(GL_TRIANGLE_STRIP));
+    {
+      YAE_OGL_11(glVertex2d(x0, y0));
+      YAE_OGL_11(glVertex2d(x0, y1));
+      YAE_OGL_11(glVertex2d(x1, y0));
+      YAE_OGL_11(glVertex2d(x1, y1));
+    }
+    YAE_OGL_11(glEnd());
+
+    if (border > 0.0)
+    {
+      YAE_OGL_11(glColor4ub(colorBorder.r(),
+                            colorBorder.g(),
+                            colorBorder.b(),
+                            colorBorder.a()));
+      YAE_OGL_11(glLineWidth(border));
+      YAE_OGL_11(glBegin(GL_LINE_LOOP));
+      {
+        YAE_OGL_11(glVertex2d(x0, y0));
+        YAE_OGL_11(glVertex2d(x0, y1));
+        YAE_OGL_11(glVertex2d(x1, y1));
+        YAE_OGL_11(glVertex2d(x1, y0));
+      }
+      YAE_OGL_11(glEnd());
+    }
+  }
+
+  //----------------------------------------------------------------
+  // Rectangle::uncache
+  //
+  void
+  Rectangle::uncache()
+  {
+    border_.uncache();
+    color_.uncache();
+    colorBorder_.uncache();
+    Item::uncache();
+  }
+
+  //----------------------------------------------------------------
+  // Rectangle::paint
+  //
+  void
+  Rectangle::paintContent() const
+  {
+    BBox bbox;
+    this->get(kPropertyBBox, bbox);
+
+    double border = border_.get();
+    const Color & color = color_.get();
+    const Color & colorBorder = colorBorder_.get();
+
+    paintRect(bbox,
+              border,
+              color,
+              colorBorder);
+  }
+
+}
